Move FNO to processors
